requirements.txt
streamlit
python-dotenv
openai
PyMuPDF
langchain
langchain-community
langchain-core
langchain-huggingface
langchain-openai
langchain-text-splitters
chromadb
sentence-transformers
